requirements.txt
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.21.12
psutil==5.9.4
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
sentry-sdk==1.14.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
torch==1.13.1
torchvision==0.14.1
typing-extensions==4.4.0
urllib3==1.26.14
wandb==0.13.9
